# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Quantitative Contract Automated Trading System API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: QCAT API
  version: "1.0"
paths:
...
  /concurrent/alerts:
    get:
      consumes:
      - application/json
      description: Get current alerts and warnings from the concurrent processing
        system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  properties:
                    alerts:
                      items:
                        type: object
                      type: array
                    count:
                      type: integer
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Get alerts
      tags:
      - Concurrent
  /concurrent/load-balancer:
    get:
      consumes:
      - application/json
      description: Get statistics and metrics for the load balancer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Get load balancer statistics
      tags:
      - Concurrent
  /concurrent/monitor:
    get:
      consumes:
      - application/json
      description: Get monitoring statistics for the concurrent processing system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Get monitor statistics
      tags:
      - Concurrent
  /concurrent/pools:
    get:
      consumes:
      - application/json
      description: Get statistics for a specific pool or all pools if no pool name
        specified
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Get pool statistics
      tags:
      - Concurrent
  /concurrent/pools/{pool_name}:
    get:
      consumes:
      - application/json
      description: Get statistics for a specific pool or all pools if no pool name
        specified
      parameters:
      - description: Pool name (optional)
        in: path
        name: pool_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Get pool statistics
      tags:
      - Concurrent
  /concurrent/task-queue:
    get:
      consumes:
      - application/json
      description: Get statistics and metrics for the task queue
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Get task queue statistics
      tags:
      - Concurrent
  /concurrent/tasks:
    post:
      consumes:
      - application/json
      description: Submit a new task to the concurrent processing system
      parameters:
      - description: Task details
        in: body
        name: request
        required: true
        schema:
          properties:
            parameters:
              type: object
            priority:
              type: integer
            task_id:
              type: string
            task_name:
              type: string
            timeout:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                data:
                  properties:
                    status:
                      type: string
                    task_id:
                      type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.Response'
      summary: Submit task
      tags:
      - Concurrent
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"